Fix/workaround filtering in queries
- refactor MicaDatasetElasticSearchQueryApi->asFilter from a big switch to a set of methods on QueryTermInterface's
- Add an implementation for TodoTerm->asFilter
- for categoric like terms, lowercase search strings if the variable is not indexed as categoric

Fixes SESIDEV-217 for now, and enables filters for generic text terms (no issue for that one)

ElasticSearch used by Opal has options to 'analyse' text fields, which means they're split at whitespace and
punctuation and lowercased and maybe stemmend and such. That is great for searching in full text where you also
want to match stuff that is like the thing you specified, it is not so great for categoric-like variables with
a fixed set of possible values. Opal currently disables the analysis for categoric variables, i.e. for
variables that have a category specified in Opal. But Ontology terms and Vocabulary terms don't have categories
set on the Opal side. The current workaround is to lowercase the search terms if the variable does not have
categories. That fixes the problem, as long as the variables contain only alphanumeric + underscore codes.