Fix query update notification to not ignore changes when a study is added or removed (LP-38).
Also some significant refactoring of the code involved to enable this.

Mica represents the query results returned from Opal as a tree, with each level representing a dimension of study, variable/term, cross
variable category. The old code compared this tree recursively to the previous result to check for significant changes and 'structure changes'.
Although these dimensions are represented in the same way structurally, they represent very different domain concepts, so abstracting over them
and using a generic recursive comparison function turned out not to be such a good idea. The new code only looks for changes in the total
matched cell, or for addition/removal of studies.