Make donorID variable not query-able (and configurable). Fix LP-17
The variable used for calculating the number of patients can no longer be used in queries. The variable is no longer visible in the list of
possible variables, and if the variable ids are fabricated into a form response the server will give an error.

The variable that is used is now configurable through the variable module.