Fix for USESI-196
The sesi_variable_from feature creates a new required field on variable nodes. The node code creates new fields on new nodes, but not on
existing nodes. The nonexisting but required field generates an error when trying to save a modified version of a variable. This change sets a
default value of 0 (no) in the field on existing variable nodes in the feature's post-enable hook.

For systems where this feature was already enabled without this fix, the code in the hook can be run manually, or the feature can be disabled
and re-enabled.