fix some issues with community content
For non-system-administrators, it is now mandatory to create community content in a
community, trying to create it outside of a community will fail.

This commit fixes some issues with the automatic filling of the community field. For
variables, the group field is hidden but it does not automatically get a value assigned
since the system uses the community of the dataset to determine access to the variable. The
group setting now no longer gets filled in if the system administrator leaves the group
setting empty on purpose.