Fix bug with variables not being shown after inheritance updates
After creating a dataset that inherits from another dataset or running 'update inherited fields' on a dataset,
the changed variables are not visible in the variable search screen. They are on the 'reorder variables' tab.
This is caused by a bug in workbench_moderation. This change works around that.

Both search_index and workbench_moderation use delayed processing. Search_api marks nodes to be indexed
and then does the actual indexing in a shutdown hook. The workbench_moderation_moderate below first sets
the current revision of the variable to unpublished because it thinks that may be an old revision. It
then registers a shutdown hook to sets the node revision back to published because workbench_moderation
thinks the current version might still be in use for other things and some non-versionable related
content may not have been updated yet. (see comments in workbench_moderation_moderate)
The result is that the shutdown hook from search_api is run first, and loads the to-be-indexed variable
from the db, which still is set to draft and is thus indexed as draft. After that the
workbench_moderation shutdown hook sets the variable to published, but then the index is stale and the
variable is not marked as change.

As a workaround we make sure the variable nodes are loaded into the entity cache before workbench_moderation touches them.
Workbench_moderation doesn't clear that cache, so search_api will load the nodes from the cache (that
contains the published version) instead of the database.

issue LP-68