Fix dataset versioning/inheritance
This commit is an extensive refactoring of the dataset inheritance and
versioning code.

- Modules dataset_versioning and dataset_inheritance have been folded
into the feature modules sesi_dataset_versioning and
sesi_dataset_inheritance. They are part of the same functional units
anyway.

- The copying/cloning of nodes and variables has been extracted into a
shared module dataset_cloning on which sesi_datset_versioning and
sesi_datset_inheritance now depend.

- Workarounds for issues with workbench_moderation have been
incorporated into dataset_cloning. One is a previously solved issue that
workbench_moderation marks a node as unpublished in the database
immediately and only sets it published in a shutdown hook. The second is
that workbench_moderation keeps all kinds of metadata on node objects to
track version changes and such, so that data is also cleared by
dataset_clone.

- Another issue was with corresponding_node_references not updating the
right node if the node being saved did not come from the database but
from a form submit. Corresponding_node_references would update a node
object loaded from the database, and then a different object constructed
form a form submit would overwrite those changes. As a fix make sure
corresponding references between datasets and variables are set
correctly but not more than once.

- Factor out updating the groups on variables and variable study
attributes into a hook in sesi_communities_and_files. Also make sure the
update is always done properly. This also means it is now impossible to
set a group different from the dataset's group settings on a variable or
study variable attribute, but they should always be in sync.

This is a fix for LP-68 and (indirectly) USESI-240