Limit header in query result screen
A very large header of a query term in the query result screen can distort the page layout and make the page hard to read. This can happen with
ICD 10 codes if a node that has many child nodes is selected. As a fix limit the header length to 80 characters.

Fixes SESIDEV-290