Change eID button position
Place the eID button on a more visible place in the side bar. This required a small css fixup to make it display properly. The eid block will
now only be shown on the register page.

The eID button from the eID button module doesn't quite work on the register page as it generates a form submit and then the validation of the
login/register form requires a username to be filled in.

Fix for USESI-283 and USESI-298