Enlarge the table column that stores query terms
For ontology queries, the query term can get very large if a lot of ontology terms are selected. (For ICD10 there are about 12000 codes.) The
terms are stored serialized in a single column. This change changes the column size from default to medium, which on MySql correspons to from
64KB to 16MB, which is enough for all ICD10 codes.

Part of SESIDEV-290