New menus
I disabled the old menus (since deleted menu items are not picked up by the feature), and created a new module that upon enabling will generate
the new menus and any missing static pages. The module only does something upon enabling if it was not enabled before, and it does not
overwrite existing pages (identified by the link alias). For new menu items it will also not create a new one if one with the same name and
parents already exist.

USESI-310