Change home screen banner icons
Part of USESI-310

Changed the text/links/icons of the homescreen banner icons. There now are only four instead of five buttons, but the texts are longer so they
need to wrap. I have adjusted the html/css so that the layout is not totally broken, but the icons/texts are still not properly centered
relative to each other, and I could not get a layout with the text next to the icons to work properly.