Code changes to get proper community content views.
This is only part of the needed changes! Still to be done: edit the appropriate feature (probably sesi_communities_and_files), and add the
og_group_ref as an indexed field to each search api index, then make sure all indexes are reindexed.

This change removes the post-processing hook in views_access_restrictions and replaces it by a hook_search_api_query_alter implementation that
adds conditions to the query to only return visible results.

The code is adapted from search_api_search-api_query_alter, please compare the implementations to see how this code was made.